+++ fhir/resources/tests/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-__fhir_version__ = "4.0.1"