+++ fhir/resources/tests/conftest.py
@@ -1,1 +1,0 @@
-pytest_plugins = ["fhir.resources.tests.fixtures"]